+++ src/InvoiceData/Settings/FormatSettingsInterface.php
@@ -13,9 +13,6 @@ interface FormatSettingsInterface
 {
     public function getFileNameFormat(): string;
 
-    // todo: split into separate interface, it should be in Pdf namespace
-    public function isForArchive(): bool;
-
     public function getInvoiceNumberFormat(): string;
 
     public function getInvoiceDateFormat(): string;