+++ src/Pdf/Settings/PdfSettingsInterface.php
@@ -1,0 +1,13 @@
+<?php
+
+/**
+ * Copyright © MB Arbatos Klubas. All rights reserved.
+ * See LICENSE file for license details.
+ */
+
+namespace FreshAdvance\Invoice\Pdf\Settings;
+
+interface PdfSettingsInterface
+{
+    public function isForArchive(): bool;
+}